% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.nback10minanimals.R
\name{process_nback_data}
\alias{process_nback_data}
\title{Process nback-10min-animals N-Back data files}
\usage{
process_nback_data(path)
}
\arguments{
\item{path}{Path to data files}
}
\description{
\code{process_breath_counting()} processes data files generated by the nback-10min-animals N-Back task
(\url{https://expfactory-experiments.github.io/nback-10min-animals}).
}
