% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.nback10minanimals.R
\name{process_nback_data_to_csv}
\alias{process_nback_data_to_csv}
\title{Convert Experiment Factory nback-10min-animals N-Back data to CSV}
\usage{
process_nback_data_to_csv(path, t, exclude)
}
\arguments{
\item{path}{Path to data files}
}
\description{
Convert Experiment Factory nback-10min-animals N-Back data to CSV
}
